+++ src/query/storages/fuse/src/operations/commit.rs
@@ -58,9 +58,9 @@ use crate::metrics::metrics_inc_commit_mutation_resolvable_conflict;
 use crate::metrics::metrics_inc_commit_mutation_retry;
 use crate::metrics::metrics_inc_commit_mutation_success;
 use crate::metrics::metrics_inc_commit_mutation_unresolvable_conflict;
-use crate::operations::commit::utils::check_label;
 use crate::operations::commit::utils::no_side_effects_in_meta_store;
 use crate::operations::mutation::AbortOperation;
+use crate::operations::util::check_duplicate_label;
 use crate::operations::AppendOperationLogEntry;
 use crate::operations::TableOperationLog;
 use crate::statistics;
@@ -214,7 +214,7 @@ impl FuseTable {
         copied_files: &Option<UpsertTableCopiedFileReq>,
         overwrite: bool,
     ) -> Result<()> {
-        if check_label(ctx.clone()).await? {
+        if check_duplicate_label(ctx.clone()).await? {
             return Ok(());
         }
         let prev = self.read_table_snapshot().await?;
@@ -778,33 +778,5 @@ mod utils {
     // check if there are any fuse table legacy options
     pub fn remove_legacy_options(table_options: &mut BTreeMap<String, String>) {
         table_options.remove(OPT_KEY_LEGACY_SNAPSHOT_LOC);
-    }
-
-    #[inline]
-    pub async fn check_label(ctx: Arc<dyn TableContext>) -> Result<bool> {
-        let duplicate_label = ctx.get_settings().get_duplicate_label().unwrap();
-        if duplicate_label.is_empty() {
-            Ok(false)
-        } else {
-            let kv_store = UserApiProvider::instance().get_meta_store_client();
-            let raw = kv_store.get_kv(&duplicate_label).await?;
-            match raw {
-                None => {
-                    let expire_at = SeqV::<()>::now_ms() / 1000 + 24 * 60 * 60 * 1000;
-                    let _ = kv_store
-                        .upsert_kv(UpsertKV {
-                            key: duplicate_label,
-                            seq: MatchSeq::Any,
-                            value: Operation::Update(1_i8.to_le_bytes().to_vec()),
-                            value_meta: Some(KVMeta {
-                                expire_at: Some(expire_at),
-                            }),
-                        })
-                        .await?;
-                    Ok(false)
-                }
-                Some(_) => Ok(true),
-            }
-        }
     }
 }
